- Refactored Identity package. Now allows for custom IdentityManager, email tasks, and authenticators.   Identity agilephp.xml configuration element and attributes are now optional - Refactored ORMFactory::loadDialect to ORMFactory::load - Comment fix to Remoting class. - Began aligning code with PEAR coding standards. - Created SessionProvider and abstracted persistence operations from the SessionScope component. Now   custom SessionProvider implementations can be created and configured using agilephp.xml. Identity   package re-integrated with SessionScope component. - Bug fix to ApplicationScope::destroy resulted in error if temp file has already been deleted - Created Mailer component - Removed FirePHP from build target since its not being used. Updated README and TODO
git-svn-id: https://agilephp.svn.sourceforge.net/svnroot/agilephp/trunk@123 9bb1911b-6204-4126-9146-82cb5eeae392

+++ src/scope/ApplicationScope.php
@@ -64,9 +64,7 @@ class ApplicationScope {
 	   * @return An instance of ApplicationScope which contains the state for the specified application
 	   * @static
 	   */
-	  public static function getInstance( $appName = null ) {
-
-	  		 self::$appName = ($appName == null) ? $_SERVER['HTTP_HOST'] : $name;
+	  public static function getInstance() {
 
 	  	     if( self::$instance == null )
 	  	         self::$instance = new self;
@@ -117,7 +115,8 @@ class ApplicationScope {
 	  public function destroy() {
 
 	  		 $this->clear();
-	  		 unlink( self::getAppTempFile() );
+	  		 if(file_exists(self::getAppTempFile()))
+	  		    unlink( self::getAppTempFile() );
 	  }
 
 	  /**
@@ -156,9 +155,10 @@ class ApplicationScope {
 	   */
 	  private static function getAppTempFile() {
 
-	  	 	  $tmp = tempnam( 'agilephp', 'agilephp-' );
- 			  $path = dirname( $tmp ) . '/agilephp-' . self::$appName;
- 			  unlink( $tmp );
+	  	 	  $tmp = tempnam('agilephp', 'agilephp-');
+ 			  $path = dirname($tmp) . '/agilephp-' . AgilePHP::getFramework()->getAppName();
+
+ 			  if(file_exists($tmp)) unlink($tmp);
 
  			  return $path;
 	  }
